Correctly specify <CR> mapping in lexima#expand example
There was a typo preventing the given example from working as provided.
This would not occur in the plugin's source code, as it uses
`lexima#string#to_mappable('<cr>')` to transform strings to prevent this
issue. However, the example does not use this function, so it needs to
be updated.

I chose not to use to use `lexima#string#to_mappable` in the example in
case this API was not meant to be public.